5.00 (Storeable) Initial commit, no tests

+++ src/ru/fizteh/fivt/students/pavel_voropaev/project/database/MultiFileTable.java
@@ -1,91 +1,100 @@
 package ru.fizteh.fivt.students.pavel_voropaev.project.database;
 
+import ru.fizteh.fivt.storage.structured.ColumnFormatException;
+import ru.fizteh.fivt.storage.structured.Storeable;
+import ru.fizteh.fivt.storage.structured.Table;
+import ru.fizteh.fivt.storage.structured.TableProvider;
 import ru.fizteh.fivt.students.pavel_voropaev.project.custom_exceptions.ContainsWrongFilesException;
 import ru.fizteh.fivt.students.pavel_voropaev.project.custom_exceptions.NullArgumentException;
-import ru.fizteh.fivt.students.pavel_voropaev.project.master.Table;
+import ru.fizteh.fivt.students.pavel_voropaev.project.custom_exceptions.TableDoesNotExistException;
 
 import java.io.*;
 import java.nio.ByteBuffer;
 import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.text.ParseException;
 import java.util.*;
 import java.util.Map.Entry;
 
 public class MultiFileTable implements Table {
-
-    private class MultiFileMap {
-        public Map<String, String> map;
-
-        public MultiFileMap() {
-            map = new HashMap<>();
-        }
-    }
-
     private static final int FOLDERS = 16;
     private static final String FOLDERS_REGEXP = "(1[0-5]|[0-9])\\.dir";
     private static final int FILES = 16;
     private static final String FILES_REGEXP = "(1[0-5]|[0-9])\\.dat";
     private static final String ENCODING = "UTF-8";
 
-    private String name;
+    private Map<String, Storeable>[] content;
+    private Map<String, Storeable> diff;
     private Path directory;
+    private String name;
+    private TableProvider parent;
     private int size;
-    private MultiFileMap[] content;
-    private Map<String, String> diff;
+    List<Class<?>> signature;
+    private boolean wasRemoved;
 
-    public MultiFileTable(Path databaseDirectory, String tableName) throws IOException {
-        name = tableName;
-        if (!Files.exists(databaseDirectory)) {
-            throw new RuntimeException(databaseDirectory.toString() + ": database doesn't exist");
+    public MultiFileTable(Path directory, String name, TableProvider parent) throws IOException {
+        if (!Files.exists(directory)) {
+            throw new RuntimeException(directory.toString() + ": database doesn't exist");
         }
-        size = 0;
-        directory = databaseDirectory.resolve(tableName);
-        content = new MultiFileMap[FOLDERS * FILES];
+
+        content = new Map[FOLDERS * FILES];
         for (int i = 0; i < FOLDERS * FILES; ++i) {
-            content[i] = new MultiFileMap();
+            content[i] = new HashMap<>();
         }
+        diff = new HashMap<>();
+        this.directory = directory.resolve(name);
+        this.name = name;
+        this.parent = parent;
+        size = 0;
+        signature = new ArrayList<>();
+        loadSignature();
+        wasRemoved = false;
 
         if (!Files.exists(directory)) {
-            try {
-                Files.createDirectory(directory);
-            } catch (IOException | SecurityException e) {
-                throw new IOException("Cannot create " + directory.getFileName());
-            }
-
-        } else {
-            load();
+            throw new IOException("Cannot read " + directory.getFileName());
         }
 
-        diff = new HashMap<>();
+        load();
     }
 
     @Override
     public String getName() {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
         return name;
     }
 
     @Override
-    public String get(String key) {
+    public Storeable get(String key) {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
         if (key == null) {
             throw new NullArgumentException("get");
         }
+
         if (diff.containsKey(key)) {
             return diff.get(key);
         }
-        return content[getPlace(key)].map.get(key);
+        return content[getPlace(key)].get(key);
     }
 
     @Override
-    public String put(String key, String value) {
+    public Storeable put(String key, Storeable value) throws ColumnFormatException {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
         if (key == null || value == null) {
             throw new NullArgumentException("put");
         }
 
+        checkStoreable(value);
         int place = getPlace(key);
-        String oldValue;
-        if (!diff.containsKey(key) && content[place].map.containsKey(key)) {
-            oldValue = content[place].map.get(key);
+        Storeable oldValue;
+        if (!diff.containsKey(key) && content[place].containsKey(key)) {
+            oldValue = content[place].get(key);
             if (oldValue.equals(value)) {
                 return oldValue;
             }
@@ -101,22 +110,25 @@ public class MultiFileTable implements Table {
     }
 
     @Override
-    public String remove(String key) {
+    public Storeable remove(String key) {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
         if (key == null) {
             throw new NullArgumentException("remove");
         }
 
         int place = getPlace(key);
-        String oldValue;
+        Storeable oldValue;
         if (!diff.containsKey(key)) {
-            if (!content[place].map.containsKey(key)) { // No such key.
+            if (!content[place].containsKey(key)) { // No such key.
                 return null;
             } else { // Key is saved on disk.
-                oldValue = content[place].map.get(key);
+                oldValue = content[place].get(key);
                 diff.put(key, null);
             }
         } else { // Some unsaved changes with this key.
-            if (!content[place].map.containsKey(key)) {
+            if (!content[place].containsKey(key)) {
                 oldValue = diff.remove(key);
             } else {
                 oldValue = diff.put(key, null);
@@ -131,22 +143,28 @@ public class MultiFileTable implements Table {
 
     @Override
     public int size() {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
         return size;
     }
 
     @Override
     public int commit() throws IOException {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
         Set<Integer> changedFiles = new HashSet<>();
-        for (Entry<String, String> entry : diff.entrySet()) {
+        for (Entry<String, Storeable> entry : diff.entrySet()) {
             String key = entry.getKey();
-            String value = entry.getValue();
+            Storeable value = entry.getValue();
             changedFiles.add(getPlace(key));
             int place = getPlace(key);
 
             if (value == null) {
-                content[place].map.remove(key);
+                content[place].remove(key);
             } else {
-                content[place].map.put(key, value);
+                content[place].put(key, value);
             }
         }
 
@@ -163,8 +181,11 @@ public class MultiFileTable implements Table {
 
     @Override
     public int rollback() {
-        for (Entry<String, String> entry : diff.entrySet()) {
-            if (content[getPlace(entry.getKey())].map.containsKey(entry.getKey())) {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
+        for (Entry<String, Storeable> entry : diff.entrySet()) {
+            if (content[getPlace(entry.getKey())].containsKey(entry.getKey())) {
                 if (entry.getValue() == null) {
                     ++size;
                 }
@@ -179,12 +200,31 @@ public class MultiFileTable implements Table {
     }
 
     @Override
-    public List<String> list() {
-        Set<String> keySet = new HashSet<>();
-        for (MultiFileMap pair : content) {
-            keySet.addAll(pair.map.keySet());
+    public int getColumnsCount() {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
         }
-        for (Entry<String, String> pair : diff.entrySet()) {
+        return signature.size();
+    }
+
+    @Override
+    public Class<?> getColumnType(int columnIndex) throws IndexOutOfBoundsException {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
+        return signature.get(columnIndex);
+    }
+
+    @Override
+    public List<String> list() {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
+        Set<String> keySet = new HashSet<>();
+        for (Map pair : content) {
+            keySet.addAll(pair.keySet());
+        }
+        for (Entry<String, Storeable> pair : diff.entrySet()) {
             if (pair.getValue() == null) {
                 keySet.remove(pair.getKey());
             } else {
@@ -198,13 +238,23 @@ public class MultiFileTable implements Table {
 
     @Override
     public int getNumberOfUncommittedChanges() {
+        if (wasRemoved) {
+            throw new TableDoesNotExistException(name);
+        }
         return diff.size();
+    }
+
+    protected void destroy() {
+        wasRemoved = true;
     }
 
     private void load() throws IOException {
         try (DirectoryStream<Path> stream = Files.newDirectoryStream(directory)) {
             for (Path entry : stream) {
                 String directoryName = entry.getFileName().toString();
+                if (directoryName.equals("signature.tsv") && Files.isRegularFile(entry)) {
+                    continue;
+                }
                 if (directoryName.matches(FOLDERS_REGEXP) && Files.isDirectory(entry)) {
                     int fileNumber = Integer.parseInt(directoryName.substring(0, directoryName.indexOf('.')));
                     if (fileNumber < FOLDERS) {
@@ -222,7 +272,7 @@ public class MultiFileTable implements Table {
     private void readDir(Path entry, int dirNum) throws IOException {
         try (DirectoryStream<Path> stream = Files.newDirectoryStream(entry)) {
             for (Path path : stream) {
-                String fileName = entry.getFileName().toString();
+                String fileName = path.getFileName().toString();
                 if (fileName.matches(FILES_REGEXP) && Files.isRegularFile(path)) {
                     int fileNumber = Integer.parseInt(fileName.substring(0, fileName.indexOf('.')));
                     if (fileNumber < FILES) {
@@ -242,17 +292,19 @@ public class MultiFileTable implements Table {
             while (stream.getFilePointer() < stream.length()) {
                 String key = readWord(stream);
                 String value = readWord(stream);
+                Storeable deserializedValue = parent.deserialize(this, value);
+
                 int hashCode = Math.abs(key.hashCode());
                 if (hashCode % FOLDERS != dirNum || hashCode / FOLDERS % FILES != fileNum) {
                     throw new ContainsWrongFilesException(directory.toString());
                 }
 
-                if (content[dirNum * FILES + fileNum].map.put(key, value) != null) { // The same key. Twice.
+                if (content[dirNum * FILES + fileNum].put(key, deserializedValue) != null) { //The same key.
                     throw new ContainsWrongFilesException(directory.toString());
                 }
                 ++size;
             }
-        } catch (EOFException e) {
+        } catch (ParseException | EOFException e) {
             throw new ContainsWrongFilesException(directory.toString());
         }
     }
@@ -275,7 +327,7 @@ public class MultiFileTable implements Table {
                 Files.createDirectory(subdirectoryPath);
             }
 
-            if (content[number].map.size() > 0) {
+            if (content[number].size() > 0) {
                 writeFile(number);
             } else {
                 if (Files.exists(getFilePath(number))) {
@@ -286,7 +338,7 @@ public class MultiFileTable implements Table {
 
         try (DirectoryStream<Path> stream = Files.newDirectoryStream(directory)) {
             for (Path entry : stream) {
-                if (entry.toFile().list().length == 0) {
+                if (Files.isDirectory(entry) && entry.toFile().list().length == 0) {
                     Files.delete(entry);
                 }
             }
@@ -299,11 +351,12 @@ public class MultiFileTable implements Table {
             Files.createFile(filePath);
         }
 
-        Set<String> keyList = content[fileNumber].map.keySet();
+        Set<String> keyList = content[fileNumber].keySet();
         try (FileOutputStream output = new FileOutputStream(filePath.toString())) {
             for (String entry : keyList) {
                 byte[] keyByte = entry.getBytes(ENCODING);
-                byte[] valueByte = content[fileNumber].map.get(entry).getBytes(ENCODING);
+                Storeable deserializedValue = content[fileNumber].get(entry);
+                byte[] valueByte = parent.serialize(this, deserializedValue).getBytes(ENCODING);
 
                 writeBytes(output, keyByte);
                 writeBytes(output, valueByte);
@@ -332,4 +385,38 @@ public class MultiFileTable implements Table {
     private Path getFilePath(int fileNumber) {
         return getDirectoryPath(fileNumber).resolve(String.valueOf(fileNumber % FOLDERS) + ".dat");
     }
+
+    private void loadSignature() throws IOException {
+        Path signaturePath = directory.resolve(Serializer.SIGNATURE_FILE_NAME);
+        if (!Files.isRegularFile(signaturePath)) {
+            throw new IOException("Wrong signature file");
+        }
+        try (Scanner scanner = new Scanner(signaturePath)) {
+            String[] types = scanner.nextLine().split("\\s+");
+            for (String entry : types) {
+                Class<?> type = Serializer.SUPPORTED_TYPES.get(entry);
+                if (type == null) {
+                    throw new IOException("Unsupported type " + entry + " in " + signaturePath);
+                }
+                signature.add(type);
+            }
+        } catch (IOException | NoSuchElementException e) {
+            throw new IOException("Unable to read signature " + signaturePath, e);
+        }
+    }
+
+    private void checkStoreable(Storeable value) throws ColumnFormatException {
+        int i = 0;
+        try {
+            for (i = 0; i < signature.size(); ++i) {
+                Object column = value.getColumnAt(i);
+                if (column != null && signature.get(i) != column.getClass()) {
+                    throw new ColumnFormatException("Expected: " + signature.get(i) + ", found: " + column.getClass());
+                }
+            }
+        } catch (IndexOutOfBoundsException e) {
+            throw new ColumnFormatException("Expected: " + signature.get(i) + ", found: <null>");
+        }
+    }
+
 }
